requirements.txt
numpy 
pandas 
matplotlib 
seaborn 
pymongo
scikit-learn
from_root
pyyaml
dill

xgboost
catboost
flask
python-dotenv